requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flask-talisman==1.0.0
Flask-WTF==1.1.1
fonttools==4.39.4
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.3
numpy-stl==3.0.1
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.5.2
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
Werkzeug==2.3.4
WTForms==3.0.1
Flask-Mail==0.9.1
PyJWT==2.8.0